# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Esta é a API de Indicadores.",
        "title": "API de Indicadores",
        "contact": {
            "name": "Suporte",
            "email": "analytics@controllab.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://victor.controllab.com:8000/api/v1"
        }
    ],
    "paths": {
        "/form": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List all forms",
                "tags": [
                    "form"
                ],
                "summary": "List Forms",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Formulario"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a form",
                "tags": [
                    "form"
                ],
                "summary": "Create Form",
                "parameters": [
                    {
                        "description": "Input",
                        "name": "form",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Formulario"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/createResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "form"
                ],
                "summary": "Update Form",
                "parameters": [
                    {
                        "description": "Input",
                        "name": "form",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Formulario"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/createResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/form/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete Form",
                "tags": [
                    "form"
                ],
                "summary": "Delete Form",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/createResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return the logged-in user's data",
                "tags": [
                    "user"
                ],
                "summary": "User",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "members": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/create": {
            "put": {
                "description": "Create new user",
                "tags": [
                    "user"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "Usuário",
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/createResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/groups": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Establishing the Relationship Between Users and Groups",
                "tags": [
                    "user"
                ],
                "summary": "Add groups",
                "parameters": [
                    {
                        "description": "Input",
                        "name": "login",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/AddGrupamentoInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/createResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Linking Users to Groups",
                "tags": [
                    "user"
                ],
                "summary": "Remove Groups",
                "parameters": [
                    {
                        "description": "Input",
                        "name": "login",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/DelGrupamentoInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/createResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "return JWT token",
                "tags": [
                    "user"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/loginRequest"
                            }
                        }
                    }
                },
                "summary": "Login",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
            "AddGrupamentoInput": {
                "type": "object",
                "required": [
                    "groups_id",
                    "user_id"
                ],
                "properties": {
                    "groups_id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                }
            },
            "DelGrupamentoInput": {
                "type": "object",
                "required": [
                    "groups_id",
                    "user_id"
                ],
                "properties": {
                    "groups_id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    }
                }
            },
            "createResponse": {
                "type": "object",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string"
                    }
                }
            },
            "loginRequest": {
                "type": "object",
                "required": [
                    "password",
                    "username"
                ],
                "properties": {
                    "password": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    }
                }
            },
...
            "Empresa": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "level": {
                        "$ref": "#/components/schemas/Nivel"
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
            "Formulario": {
                "type": "object",
                "required": [
                    "active",
                    "name"
                ],
                "properties": {
                    "active": {
                        "type": "boolean"
                    },
                    "description": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
            "Grupamento": {
                "type": "object",
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
...
            "Nivel": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nome": {
                        "type": "string"
                    }
                }
            },
...
            "User": {
                "type": "object",
                "required": [
                    "first_name",
                    "last_name",
                    "password",
                    "user"
                ],
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "enterprise": {
                        "$ref": "#/components/schemas/Empresa"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "groups": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Grupamento"
                        }
                    },
                    "id": {
                        "type": "integer"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "manager_id": {
                        "type": "integer"
                    },
                    "password": {
                        "type": "string"
                    },
                    "user": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}